+++ src/main/java/com/programmers/cafe/dto/OrderRequest.java
@@ -1,5 +1,7 @@
 package com.programmers.cafe.dto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,8 +14,13 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderRequest {
+    @NotNull
+    @Email
     private String email;
+    @NotNull
     private String address;
+    @NotNull
     private String postalCode;
+    @NotNull
     private List<ProductOrderRequest> productOrders;
 }
