+++ src/main/java/com/programmers/cafe/service/OrderService.java
@@ -2,11 +2,8 @@ package com.programmers.cafe.service;
 
 import com.programmers.cafe.dto.OrderDto;
 import com.programmers.cafe.dto.OrderFilterDto;
-import com.programmers.cafe.dto.OrderRequestDto;
-import com.programmers.cafe.dto.ProductOrderDto;
 import com.programmers.cafe.entity.Order;
 import com.programmers.cafe.entity.Product;
-import com.programmers.cafe.entity.ProductOrder;
 import com.programmers.cafe.repository.OrderRepository;
 import com.programmers.cafe.repository.ProductOrderRepository;
 import com.programmers.cafe.repository.ProductRepository;
@@ -17,7 +14,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
@@ -29,34 +25,6 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final ProductRepository productRepository;
     private final ProductOrderRepository productOrderRepository;
-
-    @Transactional
-    public Order createOrder(OrderRequestDto orderRequestDto) {
-        // 주문 객체 생성
-        Order order = new Order();
-        order.setEmail(orderRequestDto.getEmail());
-        order.setAddress(orderRequestDto.getAddress());
-        order.setPostalCode(orderRequestDto.getPostalCode());
-
-        // 주문 저장 (id가 생성되고, productOrders와 연관될 준비)
-        Order savedOrder = orderRepository.save(order);
-
-        // ProductOrder 생성 및 저장
-        for (OrderRequestDto.OrderProductDto productDto : orderRequestDto.getProducts()) {
-            Product product = productRepository.findById(productDto.getProductId())
-                    .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다."));
-
-            ProductOrder productOrder = new ProductOrder();
-            productOrder.setProduct(product);
-            productOrder.setAmount(productDto.getAmount());
-            productOrder.setOrder(savedOrder);
-
-            productOrderRepository.save(productOrder);
-        }
-
-        // 상태 및 총 금액은 엔티티의 @PrePersist로 처리됨
-        return savedOrder;
-    }
 
     public Page<OrderDto> findAllByPage(int page) {
         Page<Order> orders = orderRepository.findAll(PageRequest.of(page, 10));
@@ -112,4 +80,5 @@ public class OrderService {
         Order order = orderDto.toOrder(products);
         orderRepository.save(order);
     }
+
 }
