--- src/main/resources/templates/order_create.html
@@ -1,162 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
-    <style>
-        body {
-            background: #ddd;
-        }
-
-        .card {
-            margin: auto;
-            max-width: 950px;
-            width: 90%;
-            box-shadow: 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-            border-radius: 1rem;
-            border: transparent
-        }
-
-        .summary {
-            background-color: #ddd;
-            border-top-right-radius: 1rem;
-            border-bottom-right-radius: 1rem;
-            padding: 4vh;
-            color: rgb(65, 65, 65)
-        }
-
-        @media (max-width: 767px) {
-            .summary {
-                border-top-right-radius: unset;
-                border-bottom-left-radius: 1rem
-            }
-        }
-
-        .row {
-            margin: 0
-        }
-
-        .title b {
-            font-size: 1.5rem
-        }
-
-        .col-2,
-        .col {
-            padding: 0 1vh
-        }
-
-        img {
-            width: 3.5rem
-        }
-
-        hr {
-            margin-top: 1.25rem
-        }
-        .products {
-            width: 100%;
-        }
-        .products .price, .products .action {
-            line-height: 38px;
-        }
-        .products .action {
-            line-height: 38px;
-        }
-
-    </style>
-    <title>Hello, world!</title>
-</head>
-<body class="container-fluid">
-<div class="row justify-content-center m-4">
-    <h1 class="text-center">Grids & Circle</h1>
-</div>
-<div class="card">
-    <div class="row">
-        <div class="col-md-4 summary p-4">
-            <div>
-                <h5 class="m-0 p-0"><b>Summary</b></h5>
-            </div>
-            <hr>
-            <div class="row">
-                <h6 class="p-0">Columbia Nariñó <span class="badge bg-dark text-">2개</span></h6>
-            </div>
-            <div class="row">
-                <h6 class="p-0">Brazil Serra Do Caparaó <span class="badge bg-dark">2개</span></h6>
-            </div>
-            <div class="row">
-                <h6 class="p-0">Columbia Nariñó <span class="badge bg-dark">2개</span></h6>
-            </div>
-            <form id="order-form">
-                <div class="mb-3">
-                    <label for="email" class="form-label">이메일</label>
-                    <input type="text" class="form-control mb-1" id="email" required>
-                </div>
-                <div class="mb-3">
-                    <label for="address" class="form-label">주소</label>
-                    <input type="text" class="form-control mb-1" id="address" required>
-                </div>
-                <div class="mb-3">
-                    <label for="postalCode" class="form-label">우편번호</label>
-                    <input type="text" class="form-control" id="postalCode" required>
-                </div>
-                <div class="row pt-2 pb-2 border-top">
-                    <h5 class="col">총금액</h5>
-                    <h5 class="col text-end"></h5>
-                </div>
-                <button type="submit" class="btn btn-dark col-12">결제하기</button>
-            </form>
-        </div>
-    </div>
-</div>
-<script>
-    document.addEventListener('DOMContentLoaded', () => {
-        const orderForm = document.getElementById('order-form');
-
-        orderForm.addEventListener('submit', async (e) => {
-            e.preventDefault();
-
-            // 입력 데이터 가져오기
-            const email = document.getElementById('email').value;
-            const address = document.getElementById('address').value;
-            const postalCode = document.getElementById('postalCode').value;
-
-            // 주문 생성 API 호출
-            try {
-                const response = await fetch('http://localhost:8080/order', {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json',
-                    },
-                    body: JSON.stringify({
-                        email: email,
-                        address: address,
-                        postalCode: postalCode,
-                        // 상품Id와 총 수량은 샘플 데이터
-                        productOrders: [
-                            { productId: 1, amount: 2 },
-                            { productId: 2, amount: 1 }
-                        ],
-                    }),
-                });
-
-                // 응답 처리
-                if (response.ok) {
-                    const data = await response.json();
-                    alert('주문이 완료되었습니다.');
-                } else {
-                    const errorData = await response.json();
-                    alert('오류 발생 : ' + errorData.message);
-                }
-            } catch (error) {
-                alert('서버와의 연결에 문제가 발생했습니다.');
-                console.error(error);
-            }
-        });
-    });
-</script>
-</body>
-</html>
